requirements.txt
absl-py==2.1.0
academictorrents==2.3.3
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.1
alembic==1.14.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
bencode.py==2.0.0
bitstring==3.1.5
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.7
cloud-init==24.3.1
cloudpickle==3.1.0
colorama==0.4.4
colorlog==6.9.0
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
Cython==3.0.11
dask==2024.12.0
datasets==3.3.2
dbus-python==1.2.18
debugpy==1.8.13
decorator==5.2.1
diffprivlib==0.6.6
dill==0.3.8
distlib==0.3.9
distributed==2024.12.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
facenet==1.0.5
facenet-pytorch==2.6.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
future==0.16.0
gast==0.6.0
gdown==5.2.0
gensim==4.3.3
google-pasta==0.2.0
graphviz==0.8.4
greenlet==3.1.1
grpcio==1.68.1
h5py==3.12.1
httplib2==0.20.2
huggingface-hub==0.29.0
hyperlink==21.0.0
idna==3.3
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
incremental==21.3.0
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
libmr==0.1.9
locket==1.0.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mxnet==1.9.1
mxnet-mkl==1.6.0
namex==0.0.8
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
optuna==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.8.0
phe==1.5.0
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPubSub==3.3.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
python-version==0.0.2
pytz==2022.1
PyYAML==5.4.1
pyzmq==26.3.0
ray==2.40.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.13.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sklearn-compat==0.1.3
skorch==1.0.0
smart-open==7.1.0
sortedcontainers==2.4.0
sos==4.7.2
soupsieve==2.6
SQLAlchemy==2.0.36
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tblib==3.0.0
tenseal==0.3.16
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
timm==1.0.14
toolz==1.0.0
torch==2.2.2
torchattacks==3.5.1
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trash-cli==0.17.1.14
triton==2.2.0
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.29.2
wadllib==1.3.6
watermark==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
xkit==0.0.0
xxhash==3.5.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
zope.interface==5.4.0
